requirements.txt
pandas==2.1.4
numpy==1.26.2
pyupbit==0.2.33
requests==2.31.0
notion-client==2.2.1
slack-sdk==3.26.1
python-dotenv==1.0.0
plotly==5.18.0
sympy==1.12
schedule==1.2.0
pytz==2023.3